Dejando un lobby consumer, borrando el app de friends, implementando el unico endoint que se usaba en el userauth.
