--- backend/server/api/friends/admin.py
@@ -1,6 +1,0 @@
-from django.contrib import admin
-from .models import Friendship
-
-# Register your models here.
-
-admin.site.register(Friendship)
