+++ backend/server/api/userauth/admin.py
@@ -1,6 +1,7 @@
 from django.contrib import admin
 
 # Register your models here.
-from .models import CustomUser
+from .models import CustomUser, Friendship
 
 admin.site.register(CustomUser)
+admin.site.register(Friendship)
