+++ backend/server/api/userauth/urls.py
@@ -2,7 +2,7 @@ from django.urls import path
 from django.conf import settings
 from django.conf.urls.static import static
 from .views.auth import signup_view, login_view
-from .views.user import info_me_view, info_me_id_view, update_avatar_view, get_user_id, user_exists, user_intra_exists
+from .views.user import info_me_view, info_me_id_view, update_avatar_view, get_user_id, user_exists, user_intra_exists, user_friends_list
 from .views.token import send_csrf_token_view, validate_jwt_token_view
 from .views.googleauth import enable_2fa, disable_2fa, display_qr_code, verify_totp_code, user_2fa_setup_complete
 
@@ -14,8 +14,9 @@ urlpatterns = [
     path('user/validate-jwt/', validate_jwt_token_view, name="validate jwt token"),
 
     path('user/info-me/', info_me_view, name='info-me'),	
-    path('user/info-me-id/', info_me_id_view, name='info-me-id'),	
+    path('user/info-me-id/<int:user_id>/', info_me_id_view, name='info-me-id'),	
     path('user/update-avatar/', update_avatar_view, name='update-avatar'),
+    path('user/friendlist/', user_friends_list, name='user_friendlist'),
 	
     # BORRAR ESTA
 	path('get_user_id/<str:username>/', get_user_id, name='get_user_from_username_view'),
@@ -28,7 +29,8 @@ urlpatterns = [
     path('enable_2fa/', enable_2fa, name='enable_2fa'),
     path('disable_2fa/<int:user_id>/', disable_2fa, name='disable_2fa'),
     path('display_qr_code/<int:user_id>/', display_qr_code, name='display_qr_code'),
-	path('is_2fa_setup_complete/', user_2fa_setup_complete, name="is_2fa_setup_complete")
+	path('is_2fa_setup_complete/', user_2fa_setup_complete, name="is_2fa_setup_complete"),
+	
 ] 
 urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
 urlpatterns += static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
