+++ frontend/srcs/utils/lobby/lobby-connection-handler.js
@@ -30,7 +30,7 @@ const connectWebSocket = async () => {
 
   return new Promise((resolve, reject) => {
     lobbySocket = new WebSocket(
-      `ws://localhost:8001/ws/lobby2/?token=${authToken}`
+      `ws://localhost:8001/ws/lobby/?token=${authToken}`
     );
 
     lobbySocket.addEventListener("open", (event) => {
